feat: [유저 정보] 현재 로그인한 유저 번호 저장 - kww

+++ gather-e_android/app/src/main/java/com/kosa/gather_e/model/entity/user/UserEntity.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class UserEntity(
     @SerializedName("userSeq")
-    val userSeq: Int,
+    val userSeq: Long,
     @SerializedName("userName")
     val userName: String,
     @SerializedName("userProfileImg")
